+++ src/models/bom.ts
@@ -36,10 +36,10 @@ export class Bom {
   components: ComponentRepository
   vulnerabilities: VulnerabilityRepository
 
-  /** @see version */
+  /** @see {@link version} */
   #version: PositiveInteger = 1
 
-  /** @see serialNumber */
+  /** @see {@link serialNumber} */
   #serialNumber?: UrnUuid
 
   // Property `bomFormat` is not part of model, it is runtime information.
@@ -49,8 +49,8 @@ export class Bom {
   // The dependency graph can be normalized on render-time, no need to store it in the bom model.
 
   /**
-   * @throws {TypeError} if `op.version` is neither {@link PositiveInteger} nor `undefined`
-   * @throws {TypeError} if `op.serialNumber` is neither {@link UrnUuid} nor `undefined`
+   * @throws {@link TypeError} if `op.version` is neither {@link PositiveInteger} nor `undefined`
+   * @throws {@link TypeError} if `op.serialNumber` is neither {@link UrnUuid} nor `undefined`
    */
   constructor (op: OptionalBomProperties = {}) {
     this.metadata = op.metadata ?? new Metadata()
@@ -65,7 +65,7 @@ export class Bom {
   }
 
   /**
-   * @throws {TypeError} if value is not {@link PositiveInteger}
+   * @throws {@link TypeError} if value is not {@link PositiveInteger}
    */
   set version (value: PositiveInteger) {
     if (!isPositiveInteger(value)) {
@@ -79,7 +79,7 @@ export class Bom {
   }
 
   /**
-   * @throws {TypeError} if value is neither {@link UrnUuid} nor `undefined`
+   * @throws {@link TypeError} if value is neither {@link UrnUuid} nor `undefined`
    */
   set serialNumber (value: UrnUuid | undefined) {
     if (value !== undefined && !isUrnUuid(value)) {