+++ src/types/cpe.ts
@@ -21,7 +21,7 @@ Copyright (c) OWASP Foundation. All Rights Reserved.
  * Define the format for acceptable CPE URIs. Supports CPE 2.2 and CPE 2.3 formats.
  * Refer to {@link https://nvd.nist.gov/products/cpe} for official specification.
  *
- * @see isCPE
+ * @see {@link isCPE}
  */
 export type CPE = string