+++ HISTORY.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 * Added
   * Enums for ... (via [#...])
   * Models for ... (via [#...])
+* Docs
+  * Use [TSDoc](https://tsdoc.org/) syntax in TypeScript files, instead of [JSDoc](https://jsdoc.app/) (via [#318])
 
 ## 1.10.0 - 2023-01-28